Building in DSLR rack
The DSLR rack is bigger, and uses a different motor. I also added in the gphoto module to control the DSLR through the USB cable. Right now, it automatically saves images to the Pi, so there's a large pause (5 seconds) to allow for the transfer. Still working out the kinks on file naming and GUI